Rename the nucleon models
